Pattern Matching Algorithm Implementation
This commit adds the match algorithm adapted from the book
"The Implementation of Functional Programming Languages".

It's not yet exhaustive but it's already working.

the "toplevel" declaration is the `matchProgram` function that
receives a Program, and returns a Program but with every `Def`
replaced with a `Def` without the pattern matching (only cases).